+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ## 5.1.1
 ### Back-end
 * Add reconnection.attempts, reconnection.delay options to config - applicable for editor-server connection
+* Add sockjs config section for testing purposes